Study of Networks in Chess Games
The raw data that used is supplied from "database.lichess.com" which is a open source and free database. There are plenty of methods to download and import the data such as using personal API access token or OAuth or simply downloading the
provided chunk of data from download list which belongs to each month from January 2013.   Concerning the limitations of power in computing, the selected data belongs to all the standard games that are played in lichess online during February 2013 and contains 75321 single matches. The fil etyp ei sPGN(Portable Game Notation) and it is compressed by bzip2.  PGN is a standard plain text format for recording the chess games that is read/write by human. Hence, the firs challenge for analysing chess games graph network is translating PGN to machine language.  As the code is written in python, it is possible to use some precooked python functions and libraries such as chess.pgn and pgn2data to convert the pgn to FEN (Forsyth-Edward Notation).  FEN can describe the whole the game (at each stage) while PGN describes moves of one player.